+++ examples/Boards_Bluetooth/nRF8001/nRF8001.ino
@@ -29,9 +29,8 @@
 
 #define BLYNK_USE_DIRECT_CONNECT
 
-#include <BlynkSimpleSerialBLE.h>
+#include <BlynkSimpleBLEPeripheral.h>
 #include <BLEPeripheral.h>
-#include "BLESerial.h"
 #include <SPI.h>
 
 // You should get Auth Token in the Blynk App.
