+++ src/BlynkSimpleBLEPeripheral.h
@@ -1,0 +1,21 @@
+/**
+ * @file       BlynkSimpleBLEPeripheral.h
+ * @author     Volodymyr Shymanskyy
+ * @license    This project is released under the MIT License (MIT)
+ * @copyright  Copyright (c) 2016 Volodymyr Shymanskyy
+ * @date       Mar 2016
+ * @brief
+ *
+ */
+
+#ifndef BlynkSimpleBLEPeripheral_h
+#define BlynkSimpleBLEPeripheral_h
+
+#ifndef BLYNK_INFO_CONNECTION
+#define BLYNK_INFO_CONNECTION "BLEPeripheral"
+#endif
+
+#include <BlynkSimpleSerialBLE.h>
+#include <Adapters/BlynkBLEPeripheralSerial.h>
+
+#endif
